+++ tests/t-logLik-vcov-coef.R
@@ -9,12 +9,18 @@ fitg <- fitdist(serving, "gamma")
 logLik(fitg)
 vcov(fitg)
 coef(fitg)
+AIC(fitg)
+AIC(fitg, k = log(fitg$n)) # should give BIC
+BIC(fitg)
 
 fitg <- fitdist(serving, "gamma", method="mme")
 
 logLik(fitg)
 vcov(fitg) 
 coef(fitg)
+AIC(fitg)
+AIC(fitg, k = log(fitg$n)) # should give BIC
+BIC(fitg)
 
 # (2) Fit of a lognormal distribution to bacterial contamination data
 #
@@ -23,4 +29,7 @@ fitsf  <-  fitdistcens(smokedfish,"lnorm")
 logLik(fitsf)
 vcov(fitsf) 
 coef(fitsf)
+AIC(fitsf)
+AIC(fitsf, k = log(fitsf$n)) # should give BIC
+BIC(fitsf)